db_subscriber/requirements.txt
paho-mqtt>=2.0.0
psycopg2-binary>=2.9.0